+++ src/translations/en.ts
@@ -1,0 +1,10 @@
+export default {
+  pages: {
+    not_found_page: {
+      title: 'Page not found',
+      dont_cry: "Oops! Don't cry",
+      sth_went_wrong: 'Something went wrong',
+      go_home: 'Go Home',
+    },
+  },
+};