+++ src/translations/es.ts
@@ -1,0 +1,10 @@
+export default {
+  pages: {
+    not_found_page: {
+      title: 'Página no encontrada',
+      dont_cry: 'Oops! No llores',
+      sth_went_wrong: 'Algo salió mal',
+      go_home: 'Ir a inicio',
+    },
+  },
+};