+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
-## 29/11/2022
+## Sun 11/30/2022
+- Added i18n support. `en` and `es` are supported. Default is `en`.
+- Added language selector for the UI.
+
+## Sat 11/29/2022
 - Added Mui as styling framework.
 - Added `Silka` as custom font family.
 - Added dark mode.